employees/new: change to reusable component
